+++ packages/networks/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "networks/5.6.3";
+export declare const version = "networks/5.6.4";
 //# sourceMappingURL=_version.d.ts.map