+++ packages/networks/lib/index.js
@@ -54,7 +54,11 @@ function ethDefaultProvider(network) {
         }
         if (providers.AnkrProvider && options.ankr !== "-") {
             try {
-                providerList.push(new providers.AnkrProvider(network, options.ankr));
+                var skip = ["ropsten"];
+                var provider = new providers.AnkrProvider(network, options.ankr);
+                if (provider.network && skip.indexOf(provider.network.name) === -1) {
+                    providerList.push(provider);
+                }
             }
             catch (error) { }
         }