+++ packages/tests/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "tests/5.6.5";
+export declare const version = "tests/5.6.6";
 //# sourceMappingURL=_version.d.ts.map